move library file and update path
